+++ tests/testthat/test_nest.R
@@ -1,7 +1,8 @@
 library(test.nest)
 
+test_lintr()
 if (Sys.getenv("R_COVR") != "true") {
   test_strict()
 }
-test_lintr()
 test_regexp()
+test_spell()